backend/requirements.txt
asgiref==3.11.1
bcrypt==4.0.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
mongoengine==0.29.1
passlib==1.7.4
pillow==12.1.1
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
sqlparse==0.5.5
